Extract shared DdlScripter to Core (dedupe drifted copies)
The CREATE INDEX / CREATE TABLE scripting was duplicated in
PlanViewerControl.Schema.cs and QuerySessionControl.Schema.cs, and the two
copies had drifted: the query-session copy used a BracketName() helper for
safe identifier quoting (and richer comments) while the plan-viewer copy
used naive [{name}] interpolation that would double-bracket an already
bracketed identifier.

Consolidate onto the query-session (more-correct) version as a single,
unit-testable Core DdlScripter and point both call sites at it. The plan
viewer's schema scripting now also gets safe identifier bracketing —
identical output for normal names, correct for already-bracketed ones.

Adds DdlScripterTests covering clustered PK, nonclustered w/ include+
filter+options, columnstore, the bracket-safety behavior, and table/column
+ computed-column output. 77 -> 85 tests.
